@@ -136,7 +136,8 @@
                 },
                 save: function (flag) {
                     if (flag === 'P') {
-                        this.post.published = (new Date()).toUTCString();
+                        var utcString = new Date().toUTCString();
+                        this.post.published = new Date(utcString);
                     }
                     else if (flag === 'U') {
                         this.post.published = new Date('0001-01-01T00:00:00Z');
